requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-snowflake
apache-airflow-providers-dbt-cloud